+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - Upgrade Bevy to 0.17
+- Rename `YoetzSystemSet` -> `YoetzSystems`
 
 ## 0.4.0 - 2025-04-25
 ### Changed